requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fonttools==4.61.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.54.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
tenacity==9.1.2
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
websockets==15.0.1
